# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.ErrorData:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  response.StandardResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorData'
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
...
  usecase.OpenAPICatalogResponse:
    properties:
      dataset:
        items:
          $ref: '#/definitions/usecase.OpenAPIMetadataItem'
        type: array
      description:
        type: string
      license:
        type: string
      publisher:
        type: string
      title:
        type: string
      total_items:
        type: integer
    type: object
  usecase.OpenAPIMetadataItem:
    properties:
      access_rights:
        type: string
      description:
        type: string
      identifier:
        type: string
      issued:
        type: string
      landing_page:
        type: string
      modified:
        type: string
      publisher:
        type: string
      service_status:
        type: string
      target_audience:
        type: string
      theme:
        type: string
      title:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: bps1200@bps.go.id
    name: Tim Inovasi BPS Provinsi Sumatera Utara
    url: https://sumut.bps.go.id
  description: Backend RESTful API untuk Portal Aplikasi Terintegrasi BPS Provinsi
    Sumatera Utara.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sumut.bps.go.id
  title: Portal Aplikasi BPS Provinsi Sumatera Utara (ALUSI) API
  version: 1.2.0
paths:
...
  /admin/apps/{id}/probe:
    post:
      description: Memaksa pemeriksaan langsung (live HTTP ping) pada endpoint health
        check aplikasi (Admin)
      parameters:
      - description: ID Aplikasi
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Trigger manual probe kesehatan aplikasi
      tags:
      - Admin - Monitoring
...
  /admin/reports/analytics/export:
    get:
      description: Mengunduh file laporan rekapitulasi statistik penggunaan dan SLA
        dalam format CSV (Admin)
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Ekspor laporan analitik (CSV / Excel)
      tags:
      - Admin - Reports
  /admin/reports/catalog/export:
    get:
      description: Mengunduh file inventaris lengkap katalog aplikasi, URL, dan status
        dalam format CSV (Admin)
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Ekspor inventaris katalog aplikasi (CSV / Excel)
      tags:
      - Admin - Reports
...
  /apps/{id}/click:
    post:
      description: Mencatat riwayat klik pengguna dan menambah counter total hits
        secara non-blocking
      parameters:
      - description: ID Aplikasi
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.StandardResponse'
      summary: Catat klik / akses aplikasi (Asinkron)
      tags:
      - Interactions
  /apps/{id}/favorite:
    post:
      description: Melakukan toggle status bookmark / favorit aplikasi untuk pengguna
        yang login
      parameters:
      - description: ID Aplikasi
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Tambah / Hapus favorit pengguna
      tags:
      - Interactions
...
  /apps/{slug}/status-history:
    get:
      description: Mengambil riwayat log ping & health check uptime sebuah aplikasi
      parameters:
      - description: Slug atau ID aplikasi
        in: path
        name: slug
        required: true
        type: string
      - description: Batas jumlah riwayat (default 30)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.StandardResponse'
      summary: Riwayat status kesehatan aplikasi
      tags:
      - Monitoring
...
  /openapi/apps:
    get:
      description: Mengambil katalog aplikasi dalam format standar metadata terbuka
        Satu Data Indonesia & DCAT-AP
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/usecase.OpenAPICatalogResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      summary: Metadata Katalog Open API (Satu Data Indonesia / DCAT-AP)
      tags:
      - Open Data / SDI
  /services/status:
    get:
      description: Mengambil status operasional, SLA uptime persentase, dan latensi
        rata-rata seluruh aplikasi
      parameters:
      - description: Jumlah hari rekapitulasi (default 30)
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      summary: Ringkasan uptime seluruh layanan (Service Status)
      tags:
      - Monitoring
  /users/me/favorites:
    get:
      description: Mengambil semua aplikasi yang ditandai sebagai favorit oleh pengguna
        yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Daftar aplikasi favorit pengguna
      tags:
      - Interactions
  /users/me/recents:
    get:
      description: Mengambil daftar riwayat aplikasi yang baru saja diakses oleh pengguna
      parameters:
      - description: Batas jumlah item (default 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.StandardResponse'
      security:
      - BearerAuth: []
      summary: Riwayat aplikasi terakhir dibuka
      tags:
      - Interactions
securityDefinitions:
  BearerAuth:
    description: 'Masukkan JWT token dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"